reconfigured the backend code, implementing answers
instead of using multiple talbes to handle the courses, im using 1 table to hold all the questions. And the courses are now dynamic, so the next step is to make the answers appear. Currently, there is someissue with the individual quesiton pages, need to solve that first. And then move on and finish up the answers API too.